reduce risk of circular dependencies
